+++ src/collective/contact/core/utils.py
@@ -4,11 +4,13 @@ from collective.contact.core.content.person import IPerson
 from collective.contact.core.interfaces import IHeldPosition
 
 
-def get_gender_and_number(contacts):
+def get_gender_and_number(contacts, use_by=False, use_to=False):
     """Return gender and number of given contacts.
        Returns None if not genderable.
        Returns a 2 letters code if genderable:
        - first letter is for gender, M for "male", "F" for female;
+         --> if use_from, we prepend 'B' to gender, it will manage 'proposed by mister X';
+         --> if use_to, we prepend 'T' to gender, it will manage 'propose object to mister X'.
        - second letter is for number, S for "Singular", "P" for "Plural".
        p_contacts may be any kind of contacts, we will try to get person of it."""
     gender = None
@@ -35,5 +37,9 @@ def get_gender_and_number(contacts):
                 gender = person_gender
     res = gender
     if gender:
+        if use_by:
+            gender = 'B' + gender
+        elif use_to:
+            gender = 'T' + gender
         res = gender + (number > 1 and 'P' or 'S')
     return res